ReflectedNBTWrapper 1.3 WIP Commit - Moved ItemNBTImporter/ItemNBTExtractor to Legacy package - Library changed to API plugin - Now ReflectedNBTWrapper requires JDK to modify class on runtime - Block NBT Injector is working in progress
