+++ res/plugin.yml
@@ -1,0 +1,3 @@
+name: ReflectedNBTWrapper
+version: 1.0
+main: skywolf46.NBTUtil.v1_3.ReflectedNBTWrapper
