--- src/skywolf46/NBTUtil/v1_2R4/Exception/CollectionMismatchedException.java
@@ -1,7 +1,0 @@
-package skywolf46.NBTUtil.v1_2R4.Exception;
-
-public class CollectionMismatchedException extends RuntimeException {
-    public CollectionMismatchedException(Class bef, Class nex) {
-        super("Mismatching collection class: expected " + bef.getSimpleName() + " but " + nex.getSimpleName() + " detected");
-    }
-}
