--- src/skywolf46/NBTUtil/v1_2R4/NBTData/ReflectedNBTInteger.java
@@ -1,79 +1,0 @@
-package skywolf46.NBTUtil.v1_2R4.NBTData;
-
-import skywolf46.NBTUtil.v1_2R4.BukkitVersionUtil;
-import skywolf46.NBTUtil.v1_2R4.Interface.IReflectedNBTBase;
-
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
-
-public class ReflectedNBTInteger implements IReflectedNBTBase<Integer> {
-    private int d;
-    private static Class NBT_CLASS;
-    private static Constructor NBT_CONSTRUCTOR;
-    private static Field CONTENT_FIELD;
-
-    static {
-        try {
-            NBT_CLASS = BukkitVersionUtil.getNMSClass("NBTTagInt");
-            CONTENT_FIELD = NBT_CLASS.getDeclaredField("data");
-            CONTENT_FIELD.setAccessible(true);
-            NBT_CONSTRUCTOR = NBT_CLASS.getConstructor(Integer.TYPE);
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (NoSuchMethodException e) {
-            e.printStackTrace();
-        } catch (NoSuchFieldException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public ReflectedNBTInteger(Object o) {
-        if (!o.getClass().equals(NBT_CLASS)) {
-            return;
-        }
-        try {
-            this.d = (int) CONTENT_FIELD.get(o);
-        } catch (IllegalAccessException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public ReflectedNBTInteger() {
-        this.d = 0;
-    }
-
-    @Override
-    public Integer getValue() {
-        return d;
-    }
-
-    @Override
-    public IReflectedNBTBase<Integer> getNBTValue() {
-        return new ReflectedNBTInteger(this.d);
-    }
-
-    @Override
-    public void setValue(Integer d) {
-        this.d = d;
-    }
-
-    @Override
-    public void setNBTValue(IReflectedNBTBase<Integer> base) {
-        this.d = base.getValue();
-    }
-
-    @Override
-    public Object getNBTBase() {
-        try {
-            return NBT_CONSTRUCTOR.newInstance(d);
-        } catch (InstantiationException e) {
-            e.printStackTrace();
-        } catch (IllegalAccessException e) {
-            e.printStackTrace();
-        } catch (InvocationTargetException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-}
