--- src/skywolf46/NBTUtil/v1_2R4/NBTData/ReflectedNBTLong.java
@@ -1,83 +1,0 @@
-package skywolf46.NBTUtil.v1_2R4.NBTData;
-
-import skywolf46.NBTUtil.v1_2R4.BukkitVersionUtil;
-import skywolf46.NBTUtil.v1_2R4.Interface.IReflectedNBTBase;
-
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
-
-public class ReflectedNBTLong implements IReflectedNBTBase<Long> {
-    private long d;
-    private static Class NBT_CLASS;
-    private static Constructor NBT_CONSTRUCTOR;
-    private static Field CONTENT_FIELD;
-
-    static {
-        try {
-            NBT_CLASS = BukkitVersionUtil.getNMSClass("NBTTagLong");
-            CONTENT_FIELD = NBT_CLASS.getDeclaredField("data");
-            CONTENT_FIELD.setAccessible(true);
-            NBT_CONSTRUCTOR = NBT_CLASS.getConstructor(Long.TYPE);
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        } catch (NoSuchMethodException e) {
-            e.printStackTrace();
-        } catch (NoSuchFieldException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public ReflectedNBTLong(Object o) {
-        if (!o.getClass().equals(NBT_CLASS)) {
-            return;
-        }
-        try {
-            this.d = (long) CONTENT_FIELD.get(o);
-        } catch (IllegalAccessException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public ReflectedNBTLong() {
-        this.d = 0;
-    }
-
-    public ReflectedNBTLong(long l) {
-        this.d = l;
-    }
-
-    @Override
-    public Long getValue() {
-        return d;
-    }
-
-    @Override
-    public IReflectedNBTBase<Long> getNBTValue() {
-        return new ReflectedNBTLong(this.d);
-    }
-
-    @Override
-    public void setValue(Long d) {
-        this.d = d;
-    }
-
-    @Override
-    public void setNBTValue(IReflectedNBTBase<Long> base) {
-        this.d = base.getValue();
-    }
-
-    @Override
-    public Object getNBTBase() {
-        try {
-            return NBT_CONSTRUCTOR.newInstance(d);
-        } catch (InstantiationException e) {
-            e.printStackTrace();
-        } catch (IllegalAccessException e) {
-            e.printStackTrace();
-        } catch (InvocationTargetException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-}
