--- src/skywolf46/NBTUtil/v1_2R4/NBTIO/NBTInputStream.java
@@ -1,4 +1,0 @@
-package skywolf46.NBTUtil.v1_2R4.NBTIO;
-
-public class NBTInputStream {
-}
