+++ src/skywolf46/NBTUtil/v1_3/Exception/CollectionMismatchedException.java
@@ -1,0 +1,7 @@
+package skywolf46.NBTUtil.v1_3.Exception;
+
+public class CollectionMismatchedException extends RuntimeException {
+    public CollectionMismatchedException(Class bef, Class nex) {
+        super("Mismatching collection class: expected " + bef.getSimpleName() + " but " + nex.getSimpleName() + " detected");
+    }
+}
