+++ src/skywolf46/NBTUtil/v1_3/Exception/UndefinedNBTException.java
@@ -1,0 +1,7 @@
+package skywolf46.NBTUtil.v1_3.Exception;
+
+public class UndefinedNBTException extends RuntimeException {
+    public UndefinedNBTException(Class c) {
+        super("Class " + c.getSimpleName() + " not has nbt exchanger");
+    }
+}
