+++ src/skywolf46/NBTUtil/v1_3/Interface/INBTCompatible.java
@@ -1,0 +1,7 @@
+package skywolf46.NBTUtil.v1_3.Interface;
+
+import skywolf46.NBTUtil.v1_3.NBTData.ReflectedNBTCompound;
+
+public interface INBTCompatible {
+    ReflectedNBTCompound getNBT();
+}
