+++ src/skywolf46/NBTUtil/v1_3/Interface/IReflectedNBTBase.java
@@ -1,0 +1,13 @@
+package skywolf46.NBTUtil.v1_3.Interface;
+
+public interface IReflectedNBTBase<T> {
+    T getValue();
+
+    IReflectedNBTBase<T> getNBTValue();
+
+    void setValue(T value);
+
+    void setNBTValue(IReflectedNBTBase<T> base);
+
+    Object getNBTBase();
+}
