+++ src/skywolf46/NBTUtil/v1_3/NBTExtrator.java
@@ -1,0 +1,5 @@
+package skywolf46.NBTUtil.v1_3;
+
+public class NBTExtrator {
+
+}
