+++ src/skywolf46/NBTUtil/v1_3/NBTIO/NBTInputStream.java
@@ -1,0 +1,4 @@
+package skywolf46.NBTUtil.v1_3.NBTIO;
+
+public class NBTInputStream {
+}
